integrate mobile landing page
