+++ src/StarDustAdventures_frontend/src/pages/Landing.jsx
@@ -6,9 +6,13 @@ import GameConcept from '../components/landing/gameConcept/GameConcept';
 import Lore from '../components/landing/loreStoryline/Lore';
 import GradientCover from '../components/landing/GradientCover';
 import GameConceptM from '../components/landing-mobile/gameConcept/GameConceptM';
+import StarDustMobile from '../components/landing-mobile/star-mobile-landing/StarDustMobile';
 
-const GamePlayMechanics = lazy(() => import('../components/landing/gamePlayMechanics'));
-const MobileGameplayView = lazy(() => import('../components/landing-mobile/gameplayMechanics'));
+// const GamePlayMechanics = lazy(() => import('../components/landing/gamePlayMechanics'));
+// const MobileGameplayView = lazy(() => import('../components/landing-mobile/gameplayMechanics'));
+
+import GamePlayMechanics from '../components/landing/gamePlayMechanics';
+import MobileGameplayView from '../components/landing-mobile/gameplayMechanics'
 
 /**
  * PatternCover component dynamically renders gameplay mechanics based on screen width,
@@ -55,7 +59,7 @@ const Landing = ()=>{
   return (
     <div className="page">
       <Hero />
-      {width > 1024 ? <GameConcept /> : <GameConceptM />}
+      {width > 1024 ? <GameConcept /> : <StarDustMobile />}
       <PatternCover />
       <Footer />
     </div>
